requirements.txt
# Core dependencies
panel>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0

# Configuration and secrets management
python-dotenv>=1.0.0
cryptography>=41.0.0

# Optional dependencies for enhanced functionality
psutil>=5.9.0
requests>=2.28.0
